+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/common/base/BaseFragment.kt
@@ -9,7 +9,7 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
 import androidx.viewbinding.ViewBinding
-import com.bottari.presentation.common.component.LoadingDialog
+import com.bottari.presentation.view.common.LoadingDialog
 import com.google.android.material.snackbar.Snackbar
 
 abstract class BaseFragment<VB : ViewBinding>(
